优化

+++ sod/sod-dm/sod-dm-core/src/main/java/com/piesat/dm/core/api/impl/Gbase8a.java
@@ -90,6 +90,10 @@ public class Gbase8a extends DatabaseDclAbs {
     }
 
     public void deleteUser(String identifier) throws Exception {
+        int userNum = getUserNum(identifier);
+        if (userNum == 0) {
+            return;
+        }
         String sql = "DROP USER  " + identifier;
         stmt = connection.createStatement();
         stmt.execute(sql);
